Achse 2 implementiert, Programmauswahl in 1 Dropdown, Saubere Strukturierung in Klassen, Serial Port Verbindungsabbruch gehandelt

+++ OCT_GUI_Application/OCTProgram.cs
@@ -2,11 +2,13 @@
 {
     internal class OCTProgram
     {
+        // Programmstruktur des OCT Programms
+
         public string Name { get; set; }
         public double Axis0 { get; set; }
         public double Axis1 { get; set; }
 
-        public OCTProgram(string name, int axis0, int axis1)
+        public OCTProgram(string name, double axis0, double axis1)
         {
             Name = name;
             Axis0 = axis0;
